+++ src/main/java/org/example/financesaccountmentwebapp/models/Transaction.java
@@ -21,7 +21,7 @@ public class Transaction {
 
     private LocalDate date;
 
-    private Float amount;
+    private Double amount;
 
     private Boolean income;
 
@@ -61,11 +61,11 @@ public class Transaction {
         this.date = date;
     }
 
-    public Float getAmount() {
+    public Double getAmount() {
         return amount;
     }
 
-    public void setAmount(Float amount) {
+    public void setAmount(Double amount) {
         this.amount = amount;
     }
